# api/openapi.yaml
openapi: 3.1.0
info:
    title: 'Voting API'
    version: 1.0.0
    description: 'Secure voting system API — Express, TypeScript, Prisma. Auto-generated from JSDoc comments.'
servers:
    -
        url: 'http://localhost:4000'
        description: 'Local development'
    -
        url: 'https://api.yourdomain.com'
        description: Production
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: 'Paste an **access token** here.'
security:
    -
        bearerAuth: []
...
paths:
    /api/admin/dashboard:
        get:
            tags:
                - Admin
            summary: 'Get admin dashboard statistics'
            description: 'Retrieve comprehensive statistics and metrics for the admin dashboard'
            security:
                -
                    bearerAuth: []
            responses:
                '200':
                    description: 'Dashboard statistics retrieved successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    statistics: {type: object, properties: {totalVoters: {type: number, description: 'Total number of registered voters'}, totalElections: {type: number, description: 'Total number of elections'}, activeElections: {type: number, description: 'Number of currently active elections'}, totalBallots: {type: number, description: 'Total number of ballots cast'}, turnoutRate: {type: number, description: 'Voter turnout percentage'}}}
                                    voterStatusBreakdown: {type: object, additionalProperties: {type: number}, description: 'Breakdown of voters by status'}
                                    smsStats: {type: object, additionalProperties: {type: number}, description: 'SMS message statistics by status'}
                                    recentActivity: {type: array, items: {type: object, properties: {id: {type: string, format: uuid}, action: {type: string}, timestamp: {type: string, format: date-time}, user: {type: object, properties: {name: {type: string}, role: {type: string}}}, voter: {type: object, properties: {fullName: {type: string}}}}}}
                '401':
                    description: 'Unauthorized - Authentication required'
...